feat: Update token expiration to 1 hour for enhanced security; add news existence check before deletion in news controller

+++ backend/src/controllers/news.js
@@ -61,6 +61,11 @@ export const updateNews = async (req, res) => {
 export const removeNews = async (req, res) => {
   try {
     const { id } = req.params
+
+    const exists = await News.findOne({_id: id}) 
+
+    if(!exists) return res.json({msg: "News not found"})
+
     const data = await News.deleteOne({
       _id: id
     })
